# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Integration Swagger in Laravel with Passport Auth Documentation","description":"Implementation of Swagger with in Laravel","version":"1.0.0"},"paths":{"\/api\/login":{"post":{"tags":["auth"],"summary":"Sign in","description":"Login by email, password","operationId":"login","requestBody":{"description":"Pass user credentials","required":true,"content":{"application\/json":{"schema":{"required":["grant_type","client_id","client_secret"],"properties":{"username":{"type":"string","format":"email","example":"test@test.com"},"password":{"type":"string","format":"password","example":"secret"},"grant_type":{"type":"string","example":"password"},"client_id":{"type":"integer","example":"2"},"client_secret":{"type":"string"},"refresh_token":{"type":"string"}},"type":"object"}}}},"responses":{"201":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"data":{"type":"object","example":{"token_type":"Bearer","expires_in":900,"access_token":"*****","refresh_token":"*****"}}},"type":"object"}}}},"400":{"description":"Bad Request"}}}},"\/api\/register":{"post":{"tags":["auth"],"summary":"Sign up","operationId":"register","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}},{"name":"password_confirmation","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u041f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044c \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0437\u0430\u0440\u0435\u0433\u0438\u0441\u0442\u0440\u0438\u0440\u043e\u0432\u0430\u043d"}},"type":"object"}}}},"400":{"description":"Bad Request"}}}},"\/api\/logout":{"post":{"tags":["auth"],"summary":"Logout","description":"Logout user","operationId":"logout","responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u0423\u0441\u043f\u0435\u0448\u043d\u044b\u0439 \u0432\u044b\u0445\u043e\u0434 \u0438\u0437 \u0441\u0438\u0441\u0442\u0435\u043c\u044b"}},"type":"object"}}}},"401":{"description":"Unauthenticated","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"false"},"message":{"type":"string","example":"\u0412\u044b \u043d\u0435 \u0430\u0432\u0442\u043e\u0440\u0438\u0437\u043e\u0432\u0430\u043d\u044b"}},"type":"object"}}}}},"security":[{"bearer":[]}]}},"\/api\/check":{"get":{"tags":["auth"],"summary":"Check the User","description":"Check","operationId":"check","responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"data":{"type":"object","example":{"id":354,"name":"test","email":"test@test.com","email_verified_at":null,"balance":0,"remember_token":null,"created_at":"2021-02-24T03:30:57.000000Z","updated_at":"2021-02-24T03:30:57.000000Z"}}},"type":"object"}}}},"401":{"description":"Unauthenticated","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"false"},"message":{"type":"string","example":"\u0412\u044b \u043d\u0435 \u0430\u0432\u0442\u043e\u0440\u0438\u0437\u043e\u0432\u0430\u043d\u044b"}},"type":"object"}}}}},"security":[{"bearer":[]}]}},"\/api\/password\/email":{"post":{"tags":["auth"],"summary":"Send message to reset password","description":"Send message to reset password","operationId":"api\/password\/email","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u041c\u044b \u043e\u0442\u043f\u0440\u0430\u0432\u0438\u043b\u0438 \u043f\u0438\u0441\u044c\u043c\u043e \u043d\u0430 \u0412\u0430\u0448 \u043f\u043e\u0447\u0442\u043e\u0432\u044b\u0439 \u044f\u0449\u0438\u043a \u0441 \u0438\u043d\u0441\u0442\u0440\u0443\u043a\u0446\u0438\u0435\u0439 \u043f\u043e \u0441\u0431\u0440\u043e\u0441\u0443 \u043f\u0430\u0440\u043e\u043b\u044f!"}},"type":"object"}}}},"412":{"description":"Precondition Failed","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u041f\u043e\u0447\u0442\u0430 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d\u0430!"}},"type":"object"}}}},"400":{"description":"Validation Failed"}}}},"\/password\/reset":{"post":{"tags":["auth"],"summary":"Reset the Password","description":"Reset the password","operationId":"password\/reset","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}},{"name":"password_confirmation","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u041f\u0430\u0440\u043e\u043b\u044c \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u0438\u0437\u043c\u0435\u043d\u0435\u043d"}},"type":"object"}}}},"412":{"description":"Precondition Failed","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u042d\u0442\u043e\u0442 \u0442\u043e\u043a\u0435\u043d \u0441\u0431\u0440\u043e\u0441\u0430 \u043f\u0430\u0440\u043e\u043b\u044f \u043d\u0435\u0434\u0435\u0439\u0441\u0442\u0432\u0438\u0442\u0435\u043b\u0435\u043d."}},"type":"object"}}}},"400":{"description":"Validation Failed"}}}},"\/api\/email\/verify\/{id}":{"get":{"tags":["auth"],"summary":"Verify the User","description":"Verify the user","operationId":"verify","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}},{"name":"access_token","in":"query","required":false,"schema":{"type":"string"}},{"name":"expires","in":"query","required":true,"schema":{"type":"string"}},{"name":"hash","in":"query","required":true,"schema":{"type":"string"}},{"name":"signature","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u0412\u0430\u0448\u0430 \u0443\u0447\u0435\u0442\u043d\u0430\u044f \u0437\u0430\u043f\u0438\u0441\u044c \u0443\u0441\u043f\u0435\u0448\u043d\u043e \u043f\u043e\u0434\u0442\u0432\u0435\u0440\u0436\u0434\u0435\u043d\u0430"}},"type":"object"}}}},"410":{"description":"verify invalid","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u0423\u043a\u0430\u0437\u0430\u043d \u043d\u0435\u0434\u0435\u0439\u0441\u0442\u0432\u0438\u0442\u0435\u043b\u044c\u043d\u044b\u0439 \/ \u043f\u0440\u043e\u0441\u0440\u043e\u0447\u0435\u043d\u043d\u044b\u0439 URL"}},"type":"object"}}}}}}},"\/api\/email\/resend":{"get":{"tags":["auth"],"summary":"Send verify message to the User","description":"Send verify message to the user","operationId":"email\/resend","responses":{"200":{"description":"Success","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"true"},"message":{"type":"string","example":"\u041c\u044b \u043e\u0442\u043f\u0440\u0430\u0432\u0438\u043b\u0438 \u043f\u0438\u0441\u044c\u043c\u043e \u043d\u0430 \u0412\u0430\u0448 \u043f\u043e\u0447\u0442\u043e\u0432\u044b\u0439 \u044f\u0449\u0438\u043a \u0441 \u0438\u043d\u0441\u0442\u0440\u0443\u043a\u0446\u0438\u0435\u0439 \u043f\u043e \u043f\u043e\u0434\u0442\u0432\u0435\u0440\u0436\u0434\u0435\u043d\u0438\u044e \u043f\u043e\u0447\u0442\u044b!"}},"type":"object"}}}},"401":{"description":"Unauthenticated","content":{"application\/json":{"schema":{"properties":{"success":{"type":"boolean","example":"false"},"message":{"type":"string","example":"\u0412\u044b \u043d\u0435 \u0430\u0432\u0442\u043e\u0440\u0438\u0437\u043e\u0432\u0430\u043d\u044b"}},"type":"object"}}}}},"security":[{"Bearer":[]}]}}},"components":{"securitySchemes":{"Bearer":{"type":"http","description":"Enter JWT Bearer token **_only_**","scheme":"bearer","bearerFormat":"JWT"}}}}